Added more HUD elements, tweaks
Added a display panel to the armour and weapon readout section, and one for the keys. The inventory display now features the secondary row of items in the back of the selected item. However, I had to enlarge the panel to make it fit, so to keep if from blocking the screen, it's now hidden if you're not holding the Use key.